fastapi-google-stt/requirements.txt
fastapi
uvicorn
google-cloud-speech==2.0.0
python-dotenv